messages loading and text area dissapearing on click
